eepflash.sh: Improve compatibility with old "dd"s
In eepflash.sh the lines calling dd contain 'status=progress'.

This option is not compatible with dd 8.23, which is by default on my
setup (Linux raspberrypi 4.4.50-v7+), since the 'progress' option has
only been introduced as from dd 8.24. apt-get update doesn't update dd
to 8.24, so the script needs to be corrected.

The solution has been to replace 'status=progress' with 'status=none'
in the event that dd is found to be older then 8.24.

In addition to this (since the above generates an error), I noticed
that eepflash.sh is not robust to such errors: It locks the file but
doesn't free it in case an error occurs (The eeprom appears with 'UU'
when using i2cdetect -y 0).

The solution here consists in moving the 'closing' code before
returning in case of an error.

See: https://github.com/raspberrypi/hats/issues/57

Signed-off-by: Marc Dorval <marc.dorval32@gmail.com>