eepflash.sh: Use i2c bus 10 to avoid clash on Pi 4
This script used to create a bit-bashed I2C interface in the event
that I2C0 was not already enabled. Because the kernel knew of three
I2C interfaces already the soft interface appeared as i2c-3. On Pi
4 the highest numbered I2C interface is i2c-6, which causes the
soft interface to appear as i2c-7.

Work around the issue by giving the soft interface an explicit
bus number of 10, guaranteeing that it won't clash on any known
system.

See: https://github.com/raspberrypi/hats/issues/65

Signed-off-by: Phil Elwell <phil@raspberrypi.org>