Update designguide.md
Replaced specific part number with list of part numbers from major manufacturers and in two densities. All the parts meet the necessary specifications and severals are cheaper and more readily available than the original CAT24C32 reccomendation.

In these unprecedented times it can be difficult to source a specific part number for something as commoditized as small EEPROMs.