eepmake: add support for bank 1 GPIOs
Add support for optional configuration of GPIOs on bank 1:
Create a second gpio_map_d structure and instead of using a global variable
for the GPIO map use a local structure pointer that points to the correct
map depending on the GPIO number.

Also support drive, hysteresis and slew for bank 1. Use the new ATOM of
type 5 to store the settings in the EEPROM image file.

Finally extend eeprom_settings.txt by the new configuration settings for
bank 1.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>